Rename main -> solve
main() has a special meaning in Rust and although it's not strictly
forbidden to have main() in other modules, it seems to me a better solution to
have a different name for the day-modules to decrease the possibility of
confusion; especially when reading a single-day file. This way the reader can
understand a single day file is not a complete rust program (but can easily
made to one, by providing the main()-function).